[MISC][RL] Remove unused types
